+++ app/src/main/java/com/uestc/mymoa/ui/adapter/MainFragmentPagerAdapter.java
@@ -1,4 +1,4 @@
-package com.uestc.mymoa;
+package com.uestc.mymoa.ui.adapter;
 
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
