+++ app/src/main/java/com/uestc/mymoa/ui/adapter/PostAdapter.java
@@ -1,4 +1,4 @@
-package com.uestc.mymoa;
+package com.uestc.mymoa.ui.adapter;
 
 import android.content.Context;
 import android.support.v4.view.PagerAdapter;
@@ -7,6 +7,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
+
+import com.uestc.mymoa.R;
 
 import java.util.ArrayList;
 import java.util.HashMap;
