+++ app/src/main/java/com/uestc/mymoa/ui/fragment/ManageFragment.java
@@ -1,4 +1,4 @@
-package com.uestc.mymoa;
+package com.uestc.mymoa.ui.fragment;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -6,6 +6,8 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.uestc.mymoa.R;
 
 /**
  * Created by nothisboy on 2015/7/26.
