+++ app/src/main/res/layout/item_news_category.xml
@@ -2,7 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="200dp"
     android:layout_height="200dp"
-    android:background="@color/primary"
+    android:background="@color/colorPrimary"
     android:gravity="center"
     android:orientation="vertical">
