+++ app/src/main/res/layout/popupwindow_operation_menu_main.xml
@@ -2,7 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/primary"
+    android:background="@color/colorPrimary"
     android:orientation="vertical">
 
     <ListView
