+++ app/src/main/res/values/colors.xml
@@ -1,40 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-<<<<<<< HEAD
-  <color name="primary">#3C515C</color>
-  <color name="primary_dark">#142E3C</color>
-  <color name="accent">#40C4FF</color>
-  <color name="white">#FFFFFFFF</color>
-  <color name="black">#FF000000</color>
-  <color name="black_20">#33000000</color>
-  <color name="white_20">#33FFFFFF</color>
-  <color name="transparent">#00000000</color>
 
-  <color name="text_enabled">#FFFFFFFF</color>
-  <color name="text_disabled">#4DFFFFFF</color>
-
-  <color name="icon_enabled">#FFFFFFFF</color>
-  <color name="icon_disabled">#FF63727B</color>
-
-  <color name="medium_text_enabled">#FF000000</color>
-  <color name="medium_text_disabled">#4D000000</color>
-
-  <color name="light_blue_500">#03A9F4</color>
-  <color name="cyan_500">#00BCD4</color>
-  <color name="teal_500">#009688</color>
-  <color name="green_500">#4CAF50</color>
-  <color name="light_green_500">#8BC34A</color>
-
-  <integer-array name="custom_tab_colors">
-    <item>@color/light_blue_500</item>
-    <item>@color/cyan_500</item>
-    <item>@color/teal_500</item>
-    <item>@color/green_500</item>
-    <item>@color/light_green_500</item>
-  </integer-array>
-
-</resources>
-=======
     <color name="transparent">#00000000</color>
     <!-- 白色 -->
     <color name="white">#ffffff</color>
@@ -51,4 +17,3 @@
     <color name="text_white_87">#deffffff</color>
     <color name="text_white_54">#8affffff</color>
 </resources>
->>>>>>> 0573bf6d7173ce586f41c1e317ba426733b5622a
